Refactor drawing logic for graph in different files. The graph files were getting too large so refactor them.
